requirements-gui.txt
customtkinter
Pillow
py7zr
rarfile
